requirements.txt
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
humanfriendly==10.0
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
markupsafe==2.1.5
matplotlib==3.10.7
ml-dtypes==0.5.4
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
onnx==1.19.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.1
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1+cu130
torchsummary==1.5.1
torchvision==0.24.1+cu130
tqdm==4.67.1
typing-extensions==4.15.0
tzdata==2025.2
